application.properties configured

+++ hr/src/main/java/hu/webuni/hr/kinela/service/EmployeeService.java
@@ -3,5 +3,5 @@ package hu.webuni.hr.kinela.service;
 import hu.webuni.hr.kinela.model.Employee;
 
 public interface EmployeeService {
-	public float getPayRaisePercent(Employee employee);
+	public int getPayRaisePercent(Employee employee);
 }
